Implement extra functionality to analyze test-two, tick-if-tie
Implement the family of functions MeanDynamicsManyTrials,
FloatingPointApproximateRestPointManyTrials,
NDSolveMeanDynamicsManyTrials, and
EstimateSizeOfBasinOfAttractionOfVertexManyTrials. These can be used to
analyze both “test-all, min-if-tie” and “test-two, tick-if-tie”

These functions replace and generalize the old
MeanDynamicsTestAllMinIfTieManyTrials,
FloatingPointApproximateRestPointTestAllMinIfTieManyTrials,
NDSolveMeanDynamicsTestAllMinIfTieManyTrials and
EstimateSizeOfBasinOfAttractionOfVertexTestAllMinIfTieManyTrials

Add extra parameter omitFirstVariables to
NumericalGlobalStabilityOfInteriorRestPointLyapunov, to evaluate the
distance to the interior rest point ignoring the first variables.

Fix minor bug in internal function Nice[] within
RationalApproximateRestPoint, which appeared in later versions of
Mathematica (N[a] -> N[a, 50])

Add plenty of examples for test-two, tick-if-tie